+++ src/services/content-org/guided-courses.ts
@@ -10,9 +10,9 @@ const excludeFromGeneratedIndex: string[] = []
 
 const BASE_PATH: string = `/api/content-org`
 
-export async function enrollUserInGuidedCourse(guidedCourse) {
+export async function enrollUserInGuidedCourse(guidedCourse, { notifications_enabled = false }) {
   const url: string = `${BASE_PATH}/v1/user/guided-courses/enroll-user/${guidedCourse}`
-  return await fetchHandler(url, 'POST')
+  return await fetchHandler(url, 'POST', null, { notifications_enabled })
 }
 
 export async function unEnrollUserInGuidedCourse(guidedCourse) {
@@ -22,5 +22,25 @@ export async function unEnrollUserInGuidedCourse(guidedCourse) {
 
 export async function fetchEnrollmentPageMetadata(guidedCourse) {
   const url: string = `${BASE_PATH}/v1/user/guided-courses/enrollment/${guidedCourse}`
+  return await fetchHandler(url, 'GET')
+}
+
+export async function pinGuidedCourse(guidedCourse) {
+  const url: string = `${BASE_PATH}/v1/user/guided-courses/pin/${guidedCourse}`
+  return await fetchHandler(url, 'POST')
+}
+
+export async function unPinGuidedCourse(guidedCourse) {
+  const url: string = `${BASE_PATH}/v1/user/guided-courses/unpin/${guidedCourse}`
+  return await fetchHandler(url, 'POST')
+}
+
+export async function guidedCourses() {
+  const url: string = `${BASE_PATH}/v1/user/guided-courses/`
+  return await fetchHandler(url, 'GET')
+}
+
+export async function pinnedGuidedCourses() {
+  const url: string = `${BASE_PATH}/v1/user/guided-courses/pinned`
   return await fetchHandler(url, 'GET')
 }